Consolidate to single namespace; resume Android port work
Removes the duplicate `com.mlhealth.app` namespace that was introduced by
the merged feature/comprehensive-meal-planning branch. All deletions were
either superseded by main-namespace equivalents (supplement tracking) or
will be rebuilt from scratch against the iOS reference (meal plans,
recipes) in upcoming phases.

Adds new ExerciseDao, WaterDao, StringListConverter, and DateConverter
in the main namespace. Includes iOS↔Android file mapping reference
copied from the now-archived CrossPlatform-Sync directory.

+++ android/app/src/main/java/com/mochasmindlab/mlhealth/ui/screens/dashboard/DashboardScreen.kt
@@ -91,7 +91,7 @@ fun DashboardScreen(
             // Weight Card
             item {
                 WeightCard(
-                    currentWeight = uiState.currentWeight,
+                    currentWeight = uiState.currentWeight.toFloat(),
                     weightChange = uiState.weightChange,
                     lastUpdated = uiState.lastWeightDate
                 )
